Moving linalg module into a new crate
