+++ .gitignore
@@ -1,6 +1,8 @@
 # compiled code
 rusty-machine/target/
 rusty-machine/Cargo.lock
+rulinalg/target/
+rulinalg/Cargo.lock
 *~
 *.swp
 *.swo
