+++ .travis.yml
@@ -12,3 +12,6 @@ script:
   - cargo test --verbose
   - cargo build --features stats
   - cargo test --features stats
+  - cd ../rulinalg
+  - cargo build --verbose
+  - cargo test --verbose
