+++ rulinalg/src/error.rs
