+++ rulinalg/tests/lib.rs
@@ -1,0 +1,3 @@
+extern crate rulinalg;
+
+pub mod mat;
