+++ rusty-machine/Cargo.toml
@@ -16,4 +16,4 @@ stats = []
 [dependencies]
 num = {version = "0.1.31", default-features = false }
 rand = "0.3.14"
-matrixmultiply = "0.1.8"
+rulinalg = { path = "../rulinalg" }
