+++ rusty-machine/src/learning/lin_reg.rs
@@ -11,8 +11,8 @@
 //! ```
 //! use rusty_machine::learning::lin_reg::LinRegressor;
 //! use rusty_machine::learning::SupModel;
-//! use rusty_machine::linalg::matrix::Matrix;
-//! use rusty_machine::linalg::vector::Vector;
+//! use rusty_machine::linalg::Matrix;
+//! use rusty_machine::linalg::Vector;
 //!
 //! let inputs = Matrix::new(4,1,vec![1.0,3.0,5.0,7.0]);
 //! let targets = Vector::new(vec![1.,5.,9.,13.]);
@@ -31,8 +31,8 @@
 //! ```
 
 use learning::SupModel;
-use linalg::matrix::Matrix;
-use linalg::vector::Vector;
+use linalg::Matrix;
+use linalg::Vector;
 use learning::toolkit::cost_fn::CostFunc;
 use learning::toolkit::cost_fn::MeanSqError;
 use learning::optim::grad_desc::GradientDesc;
@@ -72,8 +72,8 @@ impl SupModel<Matrix<f64>, Vector<f64>> for LinRegressor {
     ///
     /// ```
     /// use rusty_machine::learning::lin_reg::LinRegressor;
-    /// use rusty_machine::linalg::matrix::Matrix;
-    /// use rusty_machine::linalg::vector::Vector;
+    /// use rusty_machine::linalg::Matrix;
+    /// use rusty_machine::linalg::Vector;
     /// use rusty_machine::learning::SupModel;
     ///
     /// let mut lin_mod = LinRegressor::default();
@@ -135,8 +135,8 @@ impl LinRegressor {
     /// ```
     /// use rusty_machine::learning::lin_reg::LinRegressor;
     /// use rusty_machine::learning::SupModel;
-    /// use rusty_machine::linalg::matrix::Matrix;
-    /// use rusty_machine::linalg::vector::Vector;
+    /// use rusty_machine::linalg::Matrix;
+    /// use rusty_machine::linalg::Vector;
     ///
     /// let inputs = Matrix::new(4,1,vec![1.0,3.0,5.0,7.0]);
     /// let targets = Vector::new(vec![1.,5.,9.,13.]);
