+++ rusty-machine/src/learning/optim/fmincg.rs
@@ -29,7 +29,7 @@
 //! - Length hard defaults to the max iterations.
 
 use learning::optim::{Optimizable, OptimAlgorithm};
-use linalg::vector::Vector;
+use linalg::Vector;
 
 use std::cmp;
 use std::f64;
