+++ rusty-machine/tests/lib.rs
@@ -1,10 +1,6 @@
 extern crate rusty_machine as rm;
 extern crate num as libnum;
 
-pub mod linalg {
-    mod mat;
-}
-
 pub mod learning {
     mod lin_reg;
     mod k_means;
